+++ example/App.test.tsx
@@ -1,9 +1,9 @@
 import { page } from '@vitest/browser/context';
 import { expect, test } from 'vitest';
 import { render } from 'vitest-browser-react';
-import App from './App.jsx';
+import App from './App.js';
 
-test('renders fluid layout', async () => {
+test('renders fluid grid layout', async () => {
   const app = render(<App />);
 
   const fullArea = app.getByTestId('area-full');
