+++ example/vite-env.d.ts
