+++ package.json
@@ -31,7 +31,8 @@
     "typescript": "^5.6.2",
     "vite": "^5.4.8",
     "vitest": "^2.1.1",
-    "vitest-browser-react": "^0.0.1"
+    "vitest-browser-react": "^0.0.1",
+    "zod": "^3.23.8"
   },
   "peerDependencies": {
     "tailwindcss": ">=3.4.6"
@@ -39,8 +40,5 @@
   "keywords": [
     "tailwindcss",
     "fluid-grid"
-  ],
-  "dependencies": {
-    "zod": "^3.23.8"
-  }
+  ]
 }
