--- src/preview/LayoutArea.tsx
@@ -1,24 +1,0 @@
-import { useMeasure } from 'react-use';
-
-export interface LayoutAreaProps {
-  name: string;
-  className: string;
-}
-
-export function LayoutArea({ name, className }: LayoutAreaProps) {
-  const [ref, { width }] = useMeasure<HTMLDivElement>();
-  return (
-    <div
-      ref={ref}
-      className={`overflow-clip ${className}`}
-      data-testid={`area-${name}`}
-    >
-      <div className="p-4 grid grid-flow-col gap-2 items-center w-fit">
-        <span className="uppercase tracking-wide font-bold text-sm">
-          {name}
-        </span>
-        <span className="text-sm text-cyan-300">({width}px)</span>
-      </div>
-    </div>
-  );
-}
