+++ tsconfig.json
@@ -15,6 +15,6 @@
     "noFallthroughCasesInSwitch": true,
     "types": ["@vitest/browser/providers/playwright"]
   },
-  "include": ["src"],
+  "include": ["src", "example"],
   "exclude": ["node_modules", "dist"]
 }
